feat: Add password field to editUser schema and normalize phone number in user updates
