+++ app/service/authService.py
@@ -10,16 +10,8 @@ from app.core.config.database import get_db
 from app.utils.security.hash import hash_password, verify_password
 from app.utils.security.jwt_handler import verify_refresh_token, verify_access_token, create_access_token, create_refresh_token, create_password_reset_token, verify_password_reset_token
 from app.utils.otp.sms import send_otp_sms, verify_otp_sms
+from app.utils.helper import normalize_phone_number, format_phone_for_sending
 
-def normalize_phone_number(phone: str) -> str:
-    # Strip +251, 251, or 0 at the start
-    return re.sub(r'^(?:\+251|251|0)', '', phone)
-
-def format_phone_for_sending(phone: str, use_plus_prefix=True) -> str:
-    if use_plus_prefix:
-        return f'+251{normalize_phone_number(phone)}'
-    else:
-        return f'0{normalize_phone_number(phone)}'
 
 class AuthService:
     def __init__(self, db):
